EVM Kernel: Unimplement unused functions from the SDK’s Runtime
For the WASM Runtime implemented in the EVM Node, we only implement the subset
of the `Runtime` trait that is stricly needed.

In order to keep track of what’s needed and what is not, we align the `Runtime`
wrapper implemented for the kernel with the `Runtime` implementation of the
WASM Runtime, with clear notes that re-implementing a function in the formel
means modifying the latter as well.