Bls12-381: Move the foreign archive in a sublibrary
We have two wrappers for Bls12-381: `src/lib_bls12_381`, and `tezos_crypto_rs`.
When building kernels natively to link them into the EVM Node, we get
duplicated symbols.

This is a long running problem, that pushed us to introduce `rust_deps` which
has become a bottleneck of its own as well. I think the root cause is: in a
monorepo context, where we often create dependencies between our libraries, we
cannot afford to add a foreign archive as part of a dependency of other
libraries. On the contrary, we should delay as much as possible the selection
of an archive containing the symbols we need, ideally up until the declaration
of binaries.

This approach would allow, for instance, to have each binary create its own
`rust_deps` foreign archive, containing only what it needs. A proper benefit
for this is for instance that the `octez-node` wouldn’t need to depend on
Wasmer anymore.

This changeset is a first step towards this change, made necessary by early
experiments on compiling natively Etherlink kernels. These kernels depends on
`tezos_crypto_rs`, which depends on blst the same way `lib_bls12_381` does.
This gives us the duplicated symbol errors. Here, I propose to split
`bils12_381` into two libraries: the one providing the OCaml wrapper, and the
one providing the foreign archive. Each executable necessitating the actual
blst implementation are then updated to depends on the newly introduce library.

Later, when we introduce natively compiled kernels, we will just have to drop
the dependency on the archive library in the EVM node, which will used the
function provided by `rust_deps` instead.

Similarly, removing `rust_deps` from the dependencies of libraries to limit it
to executables will allow to have as many `rust_deps` like libraries as
necessary, limiting the cost to build a single binaries (but potentially
increasing the cost of building them all, if it is done naively).