Node: Remove dependency to octez-l2-libs
The WASM interpreter and the WASM PVM should not be considered to be part of
the `octez-l2-libs` package, because otherwise, `octez-node` and other strictly
L1 binaries get a dependency to L2 libraries transitively.

To give an example, before this patch, building `octez-node` requires to build
`lib_wasmer`, while after the patch, it is no longer the case.