Manifest: Add inline test libraries as opam with-test dependencies
In a [previous changeset], we have added the possibility to specify additional
libraries for inline tests, but we fail to correctly report these new
dependencies in the Opam files. It seems that, out of luck, it was not a
problem for packaging, up until we start refining the dependency tree in
subsequent MRs.

This patch fixes the shortcoming, before it becomes a real problem.

[previous changeset]: https://gitlab.com/tezos/tezos/-/merge_requests/15247