Rust: Fix build on Arm64
Rust 1.78 and above are subjected to a [bug] leading Arm64 foreign archives to
contain x86 builtins. It looks like we are now triggering this very bug, which
prevents building on Arm64 hardware.

The simplest way to move forward is to downgrade our recommended version of
Rust to 1.77, and wait for a release containing the [fix] to be merged.

[bug]: https://github.com/rust-lang/rust/issues/125619
[fix]: https://github.com/rust-lang/rust/pull/131221

Co-authored-by: Ole Krüger <ole.kruger@trili.tech>