EVM Kernel: Opt-out of the blst dependency
This is a preparatory changeset for native execution. Executing kernels
natively requires to link the code of said kernels in the EVM Node, which
already have a dependency to blst through `tezos-crypto` (`lib-bls12-381`).

This is difficult for at least two reasons: (1) we don’t depend on the same
version of `blst` in the two cases, and (2) `lib-bls12-381` brings additional
wrappers through its bindings function. So it’s not like `tezos-crypto-rs` is a
drop-in replacement for `lib-bl12-381`.

In practice, we don’t use BLS in Etherlink, except maybe for the admin key. I
believe it is reasonable to prevent tz4 key to be admin of a test Etherlink
instance, hence this patchset.