DAL/Proto: copy module V1 of skip lists to namespace V2 to start its incremental migration
The commit mainly consists in duplicating the code, disabling unsed module
warning with [@@@warning "-32"], and commenting the content of the module.
In the next commits / MRs, we will inrementally uncomment and adapt the
code to work with Content_v2.t. Then, we will be able to plug it
(We include it instead of 'include V1'