EVM Node: Fix `callTracer` handling of `error`
Using the `callTracer` on a transaction being reverted due to the use of the
`REVERT` opcode in a `CALL` produces an invalid JSON. More precisely, the
`error` field is not a string, but an object with a single field
`invalid_utf8_string`. When looking at the content of this string, it becomes
obvious the `error` field is being filled by the tracer with the `outcome` of
the revert call.

It is not obvious why, but it is probably related to the issue [#7437].

The following patch is a little hackish. In a nutshell, we assume that the only
reason why the current implementation would produce an invalid utf8 string is
when a `CALL` is reverted, and we tweak the implementation to handle this case
by decoding the error as a revert output. The main advantage is that we don’t
modify the existing behavior for the valid string, so we shouldn’t see
regression.

[#7437]: https://gitlab.com/tezos/tezos/-/issues/7437