DAL/Proto: update_skip_list takes a [with_migration] argument
This is needed to correctly connect the last cell of the previous protocol
with the first cell of the new/current one. In fact, the new protocol
hashes the last cells of the previous protocol to put its hash as a
backpointer to the first cell of the new protocol. To use the correct (old)
hash function in the case, the [with_migration] tells at which level
protocolmigration happened