EVM Node: Start the observer’s TX pool in Forward mode on finalized view
By default, the EVM Node is observer mode starts its TX pool in `Relay` mode.
As such, when a user submits several consecutive transactions, the observer
keeps the transactions with a higher-nonce until they become valid. From the
perspective of the finalized view, it means a transaction is (at best)
submitted every 30 to 40s. This is not an acceptable UX for use cases requiring
to use the finalized view.

As a consequence, we change the behavior of the observer mode, to start the TX
pool in `Forward` mode instead in case the finalized view option is set.