EVM Node: Request to reload the kernel on upgrade fallback
When I wrote the WASM Runtime, I assumed the only time we have a kernel change
is after a reveal of the preimage. That is actually not quite true. When an
upgrade fails, the kernel actually restores its predecessor with a
`store_copy`.

This patch reflects that accordingly. Without this patch, the runtime actually
starts an infinite loop.