Tezt: replace spawn_command with get_and_check_version
We now use the same function to get the version for all executables.
This is more consistent.

It also means we do not depend on the Node module.

This also prepares a future commit where we will use the new function
to get the EVM node version (since the Evm_node module is not in
tezt/lib_tezos but in etherlink/tezt/lib, and it doesn't have get_version
anyway).

The regexp check is a bit redundant for most executables except the
octez-evm-node and the octez-node, because we later check version
equality with those executables, but it's not costly and it's easier
that way.