Tezt: one version test per executable
The main motivation is that this makes it easier to define the ~uses
in the following commit where compare_with_executable will vary.
This also results in finer-grained ~uses, which is nice.

This also makes it possible to select which executables to test
when testing locally, and to run those tests in parallel.
However, since now all tests have to query the node version,
it should overall be a bit slower in the CI (but in a negligible fashion).