Kernel/SDK: panic-hook: Use `std::eprintln` for printing on RISC-V
For WebAssembly, the panic hook will still use the `write_debug` host
function. However, for RISC-V, it will now use `std::eprintln!` to print
the panic message.

This re-establishes optimal optimisation effectiveness when on RISC-V.
Not using the host `write_debug` is a 3% performance boost. It is not
entirely clear to me as to why this is. My guess so far is that not
exposing the `write_debug` host function from the `core` crate seems
to result in better inlining.