CI: update external [hadolint] image
We update to the latest stable version of [hadolint] (2.12.1 has been left in beta).

We switched from debian to alpine as the image is smaller, which should be slightly more efficient in cost and duration. (5 MB vs 32 MB in amd64 versions)

Source: https://hub.docker.com/r/hadolint/hadolint/tags