Proto/DAL/Sc_rollups: A better interface for Request_adal_page
When requesting a page in the `Adjustable DAL` context, it's better for
kernels do not depend on the total number of shards because if that
values evolve on L1 during a protocol upgrade, and the kernel didn't
expect the modification (didn't upgrade or is not aware of the modification),
its internal logic might be broken.

Exposing a ratio `attestation_threshold_ppm` (attestation threshold parts per million)
allows being more generic.