EVM Node: Build Bifrost natively
This is a first step towards actually executing an Etherlink kernel natively.
And because we are humble folks, we stop ourselves at re-introducing Bifrost
(the last kernel of Etherlink voted by the community) source tree to the
monorepo, building it and linking it in the EVM node rust deps.

This at least demonstrates that we don’t have missing and/or duplicated symbols
anymore.