EVM Node: Execute Bifrost natively
And so the time has finally come, where we cross the new frontier and start
executing Etherlink’s kernel natively. Early results seems to indicate a
noticeable speed-up, so it is definitely a promising lead.

The scope of this MR is limited to executing Bifrost kernel natively. To do
so, we need to do three things:

  - We implement the `Runtime` trait for the `Host` type, which is the
    necessary condition to execute the `kernel_loop` function natively.
  - We modify the kernel itself, ever so slightly to make the execution
    possible.