Proto: Add stake_info in alpha_context
A way to get it from the back.

I suppose it is also an alternative way to get the total staking
weight, which is used in some places. I cannot ensure that this value
is available for other things than attesting, though, so let's keep it
like this for now.