Proto: Inline slot_owner in baking_rights_owner
The `Round.to_slot` step is still necessary, because it uses the
committee size. This is to only have a controllably limited number of
bakers that can potentially propose at a given level. This is
necessary (for now) in the baker. Since everyone will be able to
attest, this also means that it will be technically the only relevant
use of `committee_size`, changing its definition to `number of unique
rounds before looping bakers` or something.