Docker: Quick and dirty fix to restore Docker image generation
The generation of Docker images have been failing for a while now, with the
following error:

    ### Important packages version
    cargo: expected 1.77.2, got 1.78.0 ❌
    rust: expected 1.77.2, got 1.78.0 ❌
    opam: expected 2.2.1, got 2.3.0 ❌
    ocaml: 4.14.2 ✅
    python: 3.12.7 ✅
    poetry: 1.8.3 ✅

For opam, the reason is easy to track. It is because we are relying on the Edge
repository to get a more recent version of Opam compared to what is packaged in
Alpine 3.20.

For Rust, the rabbit hole was a lot deeper.

In a nutshell, up until a month ago, we were careful to sync our recommended
version of Rust with the one packaged in Alpine. We had to break this invariant
because of a regression in Rust 1.78 that were affecting some jobs running on
Arm64, and so we have [downgraded] the recommended version to 1.77.2. This
fixed the failing jobs, but this introduced the Docker generation image error.

Why, you ask? because the CI is actually not using `rustup` to get setup Rust,
and is still relying on the `rust` package of Alpine.

We tried moving to using `rustup`, but failed eventually because Rust on Alpine
is actually patched to work as expected. Specifically, Rust from `rustup` won’t
be able to generate `.so` files on Alpine, which are necessary for our build
system.

In the end, the easiest way to move forward and generate Docker images smoothly
again[^1], we propose to just stop checking the versions of Rust and Cargo
installed in the Docker images.

[downgraded]: https://gitlab.com/tezos/tezos/-/merge_requests/15343

[^1]: Fun story. If you retry a failed Docker image generation job today, the
  job will succeed. This is because the images generated previously have been
  published, and so the job just uses them and does not check the versions
  again.