EVM Node: Do not fail on `eth_estimateGas` with a high gas limit
The EVM Node is defensive when it comes to `eth_estimateGas`. It will fail
prematurely on requests with a gas parameter higher than 30M. Sadly, wallets
like MetaMask (browser extensions) seem to do that sometimes.

Since the kernel is already defensive enough (cap the gas consumption to 30M
before executing), we can safely drop this additional layer of protection.

Co-authored-by: Valentin Chaboche <valentin.chaboche@nomadic-labs.com>