EVM Node: Tell Alpine to use larger stack size for threads
On Alpine Linux, the stack size for threads is rather small (128KBytes on
recent versions). Since we are going to execute kernels natively, we actually
need more to deal with recursive calls of solidity contracts (because sadly
Rust does not have tail rec optimization, and SputnikVM uses recursion to
implement internal calls). Fortunately, Alpine Linux allows us to set the
default thread stack size at [link time].

[link time]: https://wiki.musl-libc.org/functional-differences-from-glibc.html#Thread-stack-size

So we do that by generating the links flag through `dune`. To know which
OS/distribution we are running on, we rely on the `/etc/os-release` file. To
give a few example, on Arch Linux, here is the content of my `/etc/os-release`
file.

    NAME="Arch Linux"
    PRETTY_NAME="Arch Linux"
    ID=arch
    BUILD_ID=rolling
    ANSI_COLOR="38;2;23;147;209"
    HOME_URL="https://archlinux.org/"
    DOCUMENTATION_URL="https://wiki.archlinux.org/"
    SUPPORT_URL="https://bbs.archlinux.org/"
    BUG_REPORT_URL="https://gitlab.archlinux.org/groups/archlinux/-/issues"
    PRIVACY_POLICY_URL="https://terms.archlinux.org/docs/privacy-policy/"
    LOGO=archlinux-logo

More relevant, here is the content of Alpine Linux 3.20.

    NAME="Alpine Linux"
    ID=alpine
    VERSION_ID=3.20.3
    PRETTY_NAME="Alpine Linux v3.20"
    HOME_URL="https://alpinelinux.org/"
    BUG_REPORT_URL="https://gitlab.alpinelinux.org/alpine/aports/-/issues"