EVM Node: Fix `reset` command for kernel upgrades
If a kernel upgrade was injected before the level of a reset, but activated
after, we need to set `applied_before` to `NULL`. This is overlooked in the
current version, but fortunately fixed with this patch.

In practice, the bug is not critical: it just prevents us to get the nice
upgrade event.