EVM/Node: sequencer flush uses [reset_at_level]
If we have 2 flushed blueprints, one at N and the second à N+1. We
start by resetting to finalized level N-1 to process N, then we need
to process N+1 and we also reset to N-1 (because we use the
checkpoints in the store). Instead of using the store, we can directly
use the flushed level N to know where we need to reset