EVM Node: Fix finalized level being store one L1 level late
The `Evm_context` module is quite tricky, due to its `session` type. This type
is a collection of mutable fields, that are updated as a way to avoid having to
return the context state everywhere. The drawback of this approach is how it
makes it difficult to support Irmin/SQLite transactions properly. The current
approach is arguably sub-optimal, in that any changes to the `session` is
accumulated in a `on_success` continuation when processing inputs. This is
potentially error-prone, as demonstrated by the current patch.

The `L1_l2_level_relationships` is used to associate to every L1 level

  - The head of the Layer 2 chain at the time of processing the L1 level, from
    the perspective of the node, and
  - The latest finalized blocks for the chain at this L1 level

The update of this table was done in `apply_evm_events` function, after
handling the events for the L1 level, but _before_ calling `on_success`.
While this is totally legitimate in appearance (this update can fail, and if it
happens, we want to rollback everything), the way it was done introduced a bug:
the finalized L2 level picked for the update was the one for the initial value
in the `session` type, _not_ the updated one.

By having `apply_evm_event_unsafe` returning the most up to date finalized
level, it becomes possible to fix this logical error.