Rust SDK: Do not use `store_has` in the happy path
Some (most?) function from the `Runtime` trait are calling `store_has` on the
input path before doing anything. In practice, this adds an overhead in every
execution branches, even in the ones where the return of the next host function
is enough te deduce the correct error to return.

In this patch, we propose to delay the call to `store_has` to only when it is
needed. This is the approach implemented in the native execution runtime, and
we have replayed the full history of Bifrost on Mainnet as evidence that the
semantics is correct.