EVM Node: Align `Runtime::store_delete` with the Rust SDK
`store_delete` (the host function) does not return an error if the input path
does not exist, while the Rust SDK guards the call of the host function with
`store_has`.

As far as we can tell, Etherlink’s kernel never tries to delete a path with
nothing behind it, but if it ever does, we will have a divergence. Considering
`store_delete` is not as heavily used as other host functions, the performance
penalty of the change is probably worth it to remain safe here.