Etherlink/Kernel/DAL: ignore invalid inputs instead of dropping the whole DAL slot
When parsing the content of a DAL slot, there are currently four
possible errors:
- unknown tag,
- failure to compute the length of a blueprint chunk,
- RLP deserialization error of the chunk,
- the total number of chunks for this blueprint is above the limit.

In the three first cases, we don't know where to resume parsing so we
simply drop the whole DAL slot. In the last case however, it is
possible to drop the invalid chunk and resume parsing after it. This
is what this commit does.