EVM Node: Add runtime versioning to SDK's handle breaking changes
Since we are re-implementing the `Runtime` trait from the Kernel SDK, we need
to prepare for the case where two kernels supported for native execution can
depend on SDK’s behaviors that are different (i.e., a breaking change has been
introduced in the SDK).

This patch does exactly that, and prepare for a [change] of semantics of the
`Runtime::store_value_size` function which has been introduced in the kernel
after the freeze of Bifrost.

[change]: https://gitlab.com/tezos/tezos/-/merge_requests/15897