Proto: rename consensus_threshold with the suffix _size
The meaning of consensus threshold will be more complicated moving
forward. Eventually it will mean the minimum staking power required to
reach the consensus. However we need to keep both the old and the new
definition, under the umbrella term "consensus_threshold". To that
effect, we decide to use the term "consensus_threshold_size" when it
pertains to the number of attested slots, mirroring the name
"consensus_committee_size".

Renamings are necessary also in the tests. It will be done in the next commit.