EVM Node: Fix timing advertized by the `blueprint_applied` event
The timing advertized was limited to the function modifying the
`session_state`, and excluding the actual application of the blueprint and
commit to the on-disk stores.

This patch addresses this by introducing a new helper `with_timing_f_e`, where
`f` stands for `function` (i.e., the continution called by the helper also
takes the result of the computation), and `e` stands for `error` (i.e., the
computation is in the error monad, not Lwt).