CI: limit [sccache] cache size to 2GB for [oc.build:static-<arch>-linux-binaries]
[oc.build:static-arm64-linux-binaries] has been systematically canceled by the OOM recently.

We believe it is due to the introduction of the [rust-deps-target] cache (!15802).

In order to keep this cache but to mitigate the OOM errors, we decrease the max size of the [sccache] cache.

We decrease it by 3 GB from 5 GB to 2 GB:
- [rust-deps-target] size was about 900 MB when OOM happened ([sccache] was at 5 GB)
- we want to have some buffer
- based on our observations, 2 GB should be high enough for [sccache] to work. (This size was reached one week after the caches being deleted.)

NB: given the architecture of CIAO, we also limit the [sccache] cache to 2 GB for [oc.build:static-amd64-linux-binaries].
Indeed, limiting [sccache] cache for the [arm64] job only would have required to add a bit of complexity to CIAO, which seems worse than limiting needlessly the [sccache] cache for the [amd64] job.