Images/CI: add [scripts/version.sh] in [inputs]
If some value in [version.sh] is modified, then the CI images should most likely be rebuilt.

The rebuild of Docker images is parameterised by the hash of files listed in [images/ci/inputs].
If any of the files listed changes, then the hash will change and the CI jobs [oc.docker:ci:<arch>] will rebuild the images.

Adding [scripts/version.sh] in [inputs] will thus trigger the rebuilt when any of the values hardcoded in [version.sh] changes.