Images/CI: force update of Alpine package list in [CI.runtime]
We have noticed then when rebuilding [CI.runtime] images, the latest version of alpine packages was not used.

This is probably because a cached version of the layers is used by [docker build].

Cf. e.g https://gitlab.com/tezos/tezos/-/jobs/8653365791#L106

To circumvent this we add the [apk upate] instructions when adding packages in [CI.runtime].

Another possibility would have been to modify [create_ci_images.sh] by using the [--no-cache] option in the call of [docker build], but this may involve unnecessary recomputations.