EVM Node: Sunset alternative [store_get_hash] with storage version
In order to deal with pre-Bifrost blocks, we have introduced an alternative
implementation for [store_get_hash], that relies on the upstream EVM node to
figure out what are the expected root hash for transaction receipts and
objects. We use an experimental feature flag in order to enable this somewhat
clunky host function, which is nice enough for now, but in a world where we
want to keep make it easy to replay the chain as fast as possible, we cannot
always use this behavior, _especially_ when it is no longer required.

In this patch, we propose a first step towards making this feature more
streamlined, by sunsetting the fallback mechanism to the upstream EVM node with
the storage version of the kernel. After V17, it is indeed no longer needed.