EVM Node: Allow to specify the network to join for observer mode
We introduce a new CLI argument `--network` which can be set when invoking the
EVM node in observer mode. If this argument is set, the node will perform
additional sanity check to ensure the consistency between its local state, the
upstream EVM node it is connected to, and the intent of the user.