EVM Node: Default to official sequencer endpoint for observer nodes
For Etherlink Mainnet, the default value is
`https://relay.mainnet.etherlink.com`, and for the Testnet it is
`https://relay.ghostnet.etherlink.com`. These two endpoints are operated by the
sequencer operator, and are endpoints specifically setup for other observer
nodes to connect to.

We implement this feature by tweaking the way the encoding are defined, so that
some fields are required except if `--network` is set by the caller. A nice
byproduct of this approach is that `init config` supports almost for free
setting these default value is `config.json`.