EVM Node: Default Mainnet initial kernel for observer nodes
The installer used to originate Etherlink Mainnet can be found on [Etherlink
Documentation], which also explains how to rebuild it locally. Still, it
remains cumbersome to use, compared to just embed it in the Octez EVM Node.

This is what this patch does. If a user does not trust the embedded installer,
they can use `--initial-kernel` to use their own.

We cannot provide a similar feature for the Testnet, considering its current
state. This should be doable after a regenesis, though.

[Etherlink Documentation]: https://docs.etherlink.com/network/building-kernel