EVM Node: Call `Random.self_init ()` from the binary
Libraries should not setup runtime components like the `Random` state. This is
commonly viewed as a bad practice: what if two libraries setup the same
components differently? Or, we can imagine tests wanting to set up a
reproducible `Random` state, instead of relying on `/dev/urandom`.

As a consequence, we move the `Random.self_init` call in `bin_node` main
module, as it is done for instance for the Octez node.