EVM Node: Support fetching the latest snapshot for supported network
With this patch, the minimal command to start a new Testnet node becomes

    octez-evm-node run observer \
        --data-dir foobar \
        --dont-track-rollup-node \
        --network testnet \
        --init-from-snapshot

When `--init-from-snapshot` is used in conjuction with `--network`, the EVM
node will download a recent snapshot from a trusted source. It is not yet
possible to specify another endpoint, but it should be relatively easy to do.

The implementation is (surprisingly?) straightforward: we try to load the Irmin
context from the chosen `data_dir`, and if the returned `init_status` is
`Created`, we download and import the snapshot (selected based on the network).
Once the import is done, we try to load again, and this time, we should end up
with a `Loaded` context. Hourra.

We had to move around some functions from `Evm_context` to a new `Data_dir`
module to break the circular dependency

       Snapshot
    -> Evm_context
    -> Snapshot

Other changes in snapshot utils modules are required to not use progress bar in
the `run` function. The `progress` library messes up the signal handlers, which
prevent `Lwt_exit` to work as expected. Thanks to this, once
https://gitlab.com/tezos/tezos/-/merge_requests/16107 will be merged, the
temporary directory will be removed when the program execution is interrupted
(say with `C-c`).

Apart from that, the code should not be the source of any particular surprises.