Lwt_utils_unix: `with_tempdir` handles the program exiting
Prior to this patch, a temporary directory created by `with_tempdir f` can leak
if the program exits while the execution of the continuation `f`. This is not
ideal, as `with_tempdir` is specifically being used to make sure the temporary
directory is deleted.

We propose to modify the function to support program termination, which can be
done using `Lwt_exit`. Sadly, the resulting implementation is not
straightforward. But it works.