clic: add ArgDefSwitch, either arg or switch with default
Adds a new kind of arguments that can be used as a switch only with
default value.
Because it may consume the first element after the argument
`--my-option` it's forbidden to be used as a global argument for now
because it could otherwise consume the first prefix of the command.