EVM Node: Wait for the sequencer in the flaky websocket scenario
The test is basically

    let* () = scenario sequencer in
    scenario observer

In `scenario sequencer` there are no particular synchronization with the
observer, which means we were probably starting the observer scenario before it
had the time to apply the 3rd blueprint.

We tweak the scenario to wait for the observer instead of the node under
scrutiny. If the observer is seeing a blueprint, the sequencer saw it before by
construction.