evm/tezt: fix batcher order test flakyness
The catchup was triggered when the rollup node was restarted but the
test want it to happens when the next L1 level is baked.

This fixes that by not creating enough blueprint to trigger a
blueprint.
The diff is that now the rollup node has the missing blueprints when
catching up and so can inject more at once.