Logging: Introduce `advertise-levels` for displaying event level
We propose to add a new opt-in feature for logging which makes explicit the
level of events, instead of forcing consumer to know about them.

  - For the `one-per-line` and `netstring` format, it adds the field `level`
    along with the existing `hostname`, `time_stamp`, `section` and `event`.
  - For the `pp-short` format, it changes the string prepends to the event to
    include the level in addition to the timestamp.
  - For the `pp_rfc5424` format, it adds the priority value at the very
    beginning of the log line—reusing the suggested severity number.

We believe this new feature will provide more flexibility to logs consumers.
For instance, with this, you can easily setup an alert in your system for any
`Error` or `Fatal` log emitted by your service. This is not something possible
to do today without the developer to adopt a strict naming convention or
similar workaround. It can also simplify debugging by allowing to `grep` errors
and warnings.

This is implemented by adding a new query parameter `advertise-levels` for
defining sinks, defaulting to `false` (the current behavior if it is not set
explicitly). This means that it is an opt-in change, that users need to enabled
themselves.

Here are examples of the new format with a project I am currently working on.

For `pp-short`:

    Jan 11 17:42:00.374   INFO │ worker started
    Jan 11 17:42:00.375 NOTICE │ Tx queue is ready
    Jan 11 17:42:00.449 NOTICE │ Floodgate is ready to spam network 0x1f47b
    Jan 11 17:42:00.449 NOTICE │   (base_fee_per_gas: 0.000000001)
    Jan 11 17:42:00.529   INFO │ Received blueprint 16790339
    Jan 11 17:42:00.700   INFO │ Injecting 1 transactions
    Jan 11 17:42:00.860   INFO │ Received blueprint 16790340

For the JSON constructed for `one-per-line` and `netstring` (pretty-printed for
readability):

    {
      "fd-sink-item.v0": {
        "level": "notice",
        "hostname": "vanellope",
        "time_stamp": 1736610520.317238,
        "section": ["floodgate"],
        "event": {
          "is_ready.v0": {
            "chain_id": "0x1f47b",
            "base_fee_per_gas":"1000000000"
          }
        }
      }
    }

Finally, for the `pp_rfc5424` log line:

    <005> 2025-01-12T14:22:43.632-00:00 [evm_node.dev.shutting_down_tx_pool] Stopping the tx-pool